Combine AreaLoader with Areas
