+++ game/src/main/resources/game.properties
@@ -12,6 +12,7 @@ objectsPath=./data/objects.yml
 stairsPath=./data/stairs.yml
 musicPath=./data/music-tracks.yml
 navGraph=./data/nav-graph.yml
+areaPath=./data/areas.yml
 objectDefinitionsPath=./data/definitions/object-definition-extras.yml
 npcDefinitionsPath=./data/definitions/npc-definition-extras.yml
 itemDefinitionsPath=./data/definitions/item-definition-extras.yml
